LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_touch-action-verification-expected.txt

Pointer Events touch-action attribute support

Test Description: Test will automatically check parsing behaviour of various touch-action combinations.

touch-action: basic verification


PASS default
PASS stylesheet-none
PASS explicit-auto
PASS explicit-pan-x
FAIL explicit-pan-left assert_equals: expected "auto" but got "pan-left"
FAIL explicit-pan-right assert_equals: expected "auto" but got "pan-right"
PASS explicit-pan-y
FAIL explicit-pan-up assert_equals: expected "auto" but got "pan-up"
FAIL explicit-pan-down assert_equals: expected "auto" but got "pan-down"
PASS explicit-pan-x-pan-y
PASS explicit-pan-y-pan-x
FAIL explicit-pan-left-pan-up assert_equals: expected "auto" but got "pan-left pan-up"
FAIL explicit-pan-left-pan-down assert_equals: expected "auto" but got "pan-left pan-down"
FAIL explicit-pan-right-pan-up assert_equals: expected "auto" but got "pan-right pan-up"
FAIL explicit-pan-right-pan-down assert_equals: expected "auto" but got "pan-right pan-down"
FAIL explicit-pan-up-pan-left assert_equals: expected "auto" but got "pan-left pan-up"
FAIL explicit-pan-up-pan-right assert_equals: expected "auto" but got "pan-right pan-up"
FAIL explicit-pan-down-pan-left assert_equals: expected "auto" but got "pan-left pan-down"
FAIL explicit-pan-down-pan-right assert_equals: expected "auto" but got "pan-right pan-down"
PASS explicit-manipulation
PASS explicit-none
PASS explicit-invalid-1
PASS explicit-invalid-2
PASS explicit-invalid-3
PASS explicit-invalid-4
PASS explicit-invalid-5
PASS explicit-invalid-6
PASS explicit-invalid-7
PASS explicit-invalid-8
PASS explicit-invalid-9
PASS explicit-invalid-10
PASS explicit-invalid-11
PASS explicit-invalid-12
PASS explicit-invalid-13
PASS not-inherited
PASS inherit
PASS initial
